// backend/src/common/errors/api-error-code.enum.ts
export enum ApiErrorCode {
  PHONE_ALREADY_REGISTERED = 'PHONE_ALREADY_REGISTERED',
  EMAIL_ALREADY_REGISTERED = 'EMAIL_ALREADY_REGISTERED',
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  ACCOUNT_SUSPENDED = 'ACCOUNT_SUSPENDED',
  UNAUTHORIZED = 'UNAUTHORIZED',
  INVALID_REFRESH_TOKEN = 'INVALID_REFRESH_TOKEN',
  INVALID_RESET_TOKEN = 'INVALID_RESET_TOKEN',
  INVALID_ACTIVATION_TOKEN = 'INVALID_ACTIVATION_TOKEN',
  BUSINESS_NAME_REQUIRED = 'BUSINESS_NAME_REQUIRED',
  FORBIDDEN = 'FORBIDDEN',
  NOT_FOUND = 'NOT_FOUND',
  SUBSCRIPTION_CUTOFF_PASSED = 'SUBSCRIPTION_CUTOFF_PASSED',
  SUBSCRIPTION_NOT_FOUND = 'SUBSCRIPTION_NOT_FOUND',
  SUBSCRIPTION_NOT_ACTIVE = 'SUBSCRIPTION_NOT_ACTIVE',
  DISTRIBUTOR_NOT_LIVE = 'DISTRIBUTOR_NOT_LIVE',
  DISTRIBUTOR_NOT_ACCEPTING = 'DISTRIBUTOR_NOT_ACCEPTING',
  DISTRIBUTOR_NOT_APPROVED = 'DISTRIBUTOR_NOT_APPROVED',
  DISTRIBUTOR_NOT_ELIGIBLE = 'DISTRIBUTOR_NOT_ELIGIBLE',
  DISTRIBUTOR_PROFILE_NOT_FOUND = 'DISTRIBUTOR_PROFILE_NOT_FOUND',
  CUSTOMER_PROFILE_NOT_FOUND = 'CUSTOMER_PROFILE_NOT_FOUND',
  CUSTOMER_NOT_FOUND = 'CUSTOMER_NOT_FOUND',
  CUSTOMER_NOT_LINKED = 'CUSTOMER_NOT_LINKED',
  NO_ACTIVE_PRICING = 'NO_ACTIVE_PRICING',
  PRICING_NOT_FOUND = 'PRICING_NOT_FOUND',
  PRICING_ALREADY_EXISTS = 'PRICING_ALREADY_EXISTS',
  INVALID_DELIVERY_SLOT = 'INVALID_DELIVERY_SLOT',
  DELIVERY_SLOT_NOT_FOUND = 'DELIVERY_SLOT_NOT_FOUND',
  PRODUCT_NOT_AVAILABLE = 'PRODUCT_NOT_AVAILABLE',
  PRODUCT_NOT_FOUND = 'PRODUCT_NOT_FOUND',
  PRODUCT_SKU_EXISTS = 'PRODUCT_SKU_EXISTS',
  PRODUCT_IN_USE = 'PRODUCT_IN_USE',
  PRODUCT_FORBIDDEN = 'PRODUCT_FORBIDDEN',
  GO_LIVE_NOT_APPROVED = 'GO_LIVE_NOT_APPROVED',
  GO_LIVE_NOT_READY = 'GO_LIVE_NOT_READY',
  ADDRESS_URBAN_LINE_REQUIRED = 'ADDRESS_URBAN_LINE_REQUIRED',
  ADDRESS_CITY_REQUIRED = 'ADDRESS_CITY_REQUIRED',
  ADDRESS_PINCODE_OR_MAP_REQUIRED = 'ADDRESS_PINCODE_OR_MAP_REQUIRED',
  ADDRESS_VILLAGE_REQUIRED = 'ADDRESS_VILLAGE_REQUIRED',
  ADDRESS_DISTRICT_REQUIRED = 'ADDRESS_DISTRICT_REQUIRED',
  ADDRESS_STATE_REQUIRED = 'ADDRESS_STATE_REQUIRED',
  ADDRESS_RURAL_LOCATION_REQUIRED = 'ADDRESS_RURAL_LOCATION_REQUIRED',
  IDENTITY_DOCUMENT_REQUIRED = 'IDENTITY_DOCUMENT_REQUIRED',
  IDENTITY_DOCUMENT_LIMIT = 'IDENTITY_DOCUMENT_LIMIT',
  IDENTITY_DOCUMENT_INVALID_TYPE = 'IDENTITY_DOCUMENT_INVALID_TYPE',
  IDENTITY_DOCUMENT_TOO_LARGE = 'IDENTITY_DOCUMENT_TOO_LARGE',
  IDENTITY_DOCUMENT_NOT_FOUND = 'IDENTITY_DOCUMENT_NOT_FOUND',
  IDENTITY_NAME_MISMATCH = 'IDENTITY_NAME_MISMATCH',
  IDENTITY_NOT_VERIFIED = 'IDENTITY_NOT_VERIFIED',
  GENERIC = 'GENERIC',
}